# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Telegram Drive API",
    "description": "HTTP API for Telegram-backed object storage. Headless (Docker) exposes the full surface on one port (1334): legacy /upload, /metrics, WebDAV, static Web UI, and /d/* share downloads. Desktop optional REST API (127.0.0.1:8550) is a subset — see docs/DESKTOP-API.md. Authenticate with X-API-Key or X-Access-Pwd (Web admin / desktop local access password).",
    "version": "3.0.0-beta"
  },
  "servers": [{ "url": "/", "description": "Current host" }],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "AccessPwdAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Access-Pwd",
        "description": "Web admin console password (ACCESS_PWD)"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/shares": {
      "get": {
        "tags": ["Shares"],
        "summary": "List share links",
        "security": [{ "AccessPwdAuth": [] }, { "ApiKeyAuth": [] }],
        "responses": { "200": { "description": "Share list" } }
      },
      "post": {
        "tags": ["Shares"],
        "summary": "Create share link",
        "security": [{ "AccessPwdAuth": [] }, { "ApiKeyAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["message_id", "file_name", "file_size"],
                "properties": {
                  "folder_id": { "type": "integer", "nullable": true },
                  "message_id": { "type": "integer" },
                  "file_name": { "type": "string" },
                  "file_size": { "type": "integer" },
                  "password": { "type": "string" },
                  "expiry_hours": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Share created" } }
      }
    },
    "/api/v1/shares/{id}": {
      "delete": {
        "tags": ["Shares"],
        "summary": "Revoke share link",
        "security": [{ "AccessPwdAuth": [] }, { "ApiKeyAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Revoked" } }
      }
    },
...
    "/d/{token}": {
      "get": {
        "tags": ["Shares"],
        "summary": "Download shared file by token",
        "parameters": [
          { "name": "token", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Binary stream or password prompt" } }
      }
    },
    "/d/{token}/verify": {
      "post": {
        "tags": ["Shares"],
        "summary": "Verify share password",
        "parameters": [
          { "name": "token", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "password": { "type": "string" } }
              }
            }
          }
        },
        "responses": { "200": { "description": "Access granted" } }
      }
    },
...
  }
}